requirements.txt
berserk
pandas
pyyaml
matplotlib
ipykernel
numpy